Refactor build to be more Python-like
This ports over some of the fixes made to wolfCrypt. Some of this will
be pre-requisites for Windows support.

Build recursion fixes:
* Don't import module being built in setup.py
* Don't build the C code until we are doing binary dist or install
* Tox tests bdist_wheel instead of source

Other fixes:
* Separate out version to separate file for modifying
* Update copyright dates
* Unification of _build_ffi and _build_wolfssl
