+++ tox.ini
@@ -2,6 +2,7 @@
 envlist = py3
 
 [testenv]
+wheel = true
 setenv =
     PYTHONPATH = {toxinidir}:{toxinidir}/wolfssl-py
