+++ wolfssl/_version.py
@@ -1,0 +1,11 @@
+# When bumping the C library version, reset the POST count to 0
+
+__wolfssl_version__ = "v5.1.1-stable"
+
+# We're using implicit post releases [PEP 440] to bump package version
+# while maintaining the C library version intact for better reference.
+# https://www.python.org/dev/peps/pep-0440/#implicit-post-releases
+#
+# MAJOR.MINOR.BUILD-POST
+
+__version__ = "5.1.1-0"
