+++ wolfssl/exceptions.py
@@ -2,7 +2,7 @@
 #
 # exceptions.py
 #
-# Copyright (C) 2006-2020 wolfSSL Inc.
+# Copyright (C) 2006-2022 wolfSSL Inc.
 #
 # This file is part of wolfSSL. (formerly known as CyaSSL)
 #
